Creating StartExpiredInspection style file
